+++ src/components/manage/Blocks/HeroImageLeft/View.jsx
@@ -5,7 +5,7 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import { flattenToAppURL } from '@plone/volto/helpers';
+import Image from '@plone/volto/components/theme/Image/Image';
 
 /**
  * View image block class.
@@ -15,14 +15,7 @@ import { flattenToAppURL } from '@plone/volto/helpers';
 const View = ({ data }) => (
   <div className="block hero">
     <div className="block-inner-wrapper">
-      {data.url && (
-        <img
-          src={`${flattenToAppURL(data.url)}/@@images/image`}
-          alt=""
-          className="hero-image"
-          loading="lazy"
-        />
-      )}
+      {data.url && <Image image={data.url} className="hero-image" />}
       <div className="hero-body">
         {data.title && <h1>{data.title}</h1>}
         {data.description && <p>{data.description}</p>}