+++ src/helpers/index.js
@@ -73,6 +73,8 @@ export {
   parseDateTime,
   normalizeLanguageName,
 } from '@plone/volto/helpers/Utils/Utils';
+
+export { getImageAttributes } from './Image/Image';
 export { messages } from './MessageLabels/MessageLabels';
 export {
   withBlockSchemaEnhancer,